AngryMaid/requirements.txt
pygame~=2.6.1
numpy~=2.0.2
pillow~=11.1.0